orchestration/requirements.txt
dagster==1.8.9
dagster-webserver==1.8.9
dbt-core==1.8.7
dbt-duckdb==1.8.1
python-dotenv==1.0.1
